// plugins/release-manager-as-a-service/src/types/types.ts
export interface Project {
  /** A unique (in the context of RMaaS) project name */
  name: string;

  /** GitHub details */
  github: {
    /**
     * Repository's organization
     *
     * @example erikengervall
     */
    org: string;
    /**
     * Repository's name
     *
     * @example dockest
     */
    repo: string;
  };

  /** Slack details */
  slack?: {
    /** Relevant slack channel for the project */
    channel: string;
    /** Link to slack channel */
    link?: string;
  };

  /**
   * Declares the versioning strategy of the project
   *
   * semver: `1.2.3` (major.minor.patch)
   * calver: `2020.01.01_0` (YYYY.0M.0D_patch)
   *
   * Default: false
   */
  versioningStrategy: 'calver' | 'semver';
}
// ...
export interface GhCreateReleaseResponse {
  /** @example 'https://api.github.com/repos/octocat/Hello-World/releases/1'; */
  url: string;
  /** @example 'https://github.com/octocat/Hello-World/releases/v1.0.0'; */
  html_url: string;
  /** @example 'https://api.github.com/repos/octocat/Hello-World/releases/1/assets'; */
  assets_url: string;
  /** @example 'https://uploads.github.com/repos/octocat/Hello-World/releases/1/assets{?name,label}'; */
  upload_url: string;
  /** @example 'https://api.github.com/repos/octocat/Hello-World/tarball/v1.0.0'; */
  tarball_url: string;
  /** @example 'https://api.github.com/repos/octocat/Hello-World/zipball/v1.0.0'; */
  zipball_url: string;
  id: number;
  /** @example 'MDc6UmVsZWFzZTE='; */
  node_id: string;
  /** @example 'v1.0.0'; */
  tag_name: string;
  /** @example 'master'; */
  target_commitish: string;
  /** @example 'v1.0.0'; */
  name: string;
  /** @example 'Description of the release'; */
  body: string;
  draft: boolean;
  prerelease: boolean;
  /** @example '2013-02-27T19:35:32Z'; */
  created_at: string;
  /** @example '2013-02-27T19:35:32Z'; */
  published_at: string;
  author: {
    /** @example 'octocat'; */
    login: string;
    id: number;
    /** @example 'MDQ6VXNlcjE='; */
    node_id: string;
    /** @example 'https://github.com/images/error/octocat_happy.gif'; */
    avatar_url: string;
    /** @example ''; */
    gravatar_id: string;
    /** @example 'https://api.github.com/users/octocat'; */
    url: string;
    /** @example 'https://github.com/octocat'; */
    html_url: string;
    /** @example 'https://api.github.com/users/octocat/followers'; */
    followers_url: string;
    /** @example 'https://api.github.com/users/octocat/following{/other_user}'; */
    following_url: string;
    /** @example 'https://api.github.com/users/octocat/gists{/gist_id}'; */
    gists_url: string;
    /** @example 'https://api.github.com/users/octocat/starred{/owner}{/repo}'; */
    starred_url: string;
    /** @example 'https://api.github.com/users/octocat/subscriptions'; */
    subscriptions_url: string;
    /** @example 'https://api.github.com/users/octocat/orgs'; */
    organizations_url: string;
    /** @example 'https://api.github.com/users/octocat/repos'; */
    repos_url: string;
    /** @example 'https://api.github.com/users/octocat/events{/privacy}'; */
    events_url: string;
    /** @example 'https://api.github.com/users/octocat/received_events'; */
    received_events_url: string;
    /** @example 'User'; */
    type: string;
    site_admin: boolean;
  };
  assets: any[];
}
// ...
export interface GhGetBranchResponse {
  /** @example 'master'; */
  name: string;
  commit: {
    /** @example '7fd1a60b01f91b314f59955a4e4d4e80d8edf11d'; */
    sha: string;
    /** @example 'MDY6Q29tbWl0N2ZkMWE2MGIwMWY5MWIzMTRmNTk5NTVhNGU0ZDRlODBkOGVkZjExZA=='; */
    node_id: string;
    commit: {
      author: {
        /** @example 'The Octocat'; */
        name: string;
        /** @example '2012-03-06T15:06:50-08:00'; */
        date: string;
        /** @example 'octocat@nowhere.com'; */
        email: string;
      };
      /** @example 'https://api.github.com/repos/octocat/Hello-World/git/commits/7fd1a60b01f91b314f59955a4e4d4e80d8edf11d'; */
      url: string;
      /** @example 'Merge pull request #6 from Spaceghost/patch-1\n\nNew line at end of file.'; */
      message: string;
      tree: {
        /** @example 'b4eecafa9be2f2006ce1b709d6857b07069b4608'; */
        sha: string;
        /** @example 'https://api.github.com/repos/octocat/Hello-World/git/trees/b4eecafa9be2f2006ce1b709d6857b07069b4608'; */
        url: string;
      };
      committer: {
        /** @example 'The Octocat'; */
        name: string;
        /** @example '2012-03-06T15:06:50-08:00'; */
        date: string;
        /** @example 'octocat@nowhere.com'; */
        email: string;
      };
      verification: {
        verified: boolean;
        /** @example 'unsigned'; */
        reason: string;
        signature: null | any;
        payload: null | any;
      };
    };
    author: {
      /** @example ''; */
      gravatar_id: string;
      /** @example 'https://secure.gravatar.com/avatar/7ad39074b0584bc555d0417ae3e7d974?d=https://a248.e.akamai.net/assets.github.com%2Fimages%2Fgravatars%2Fgravatar-140.png'; */
      avatar_url: string;
      /** @example 'https://api.github.com/users/octocat'; */
      url: string;
      id: number;
      /** @example 'octocat'; */
      login: string;
    };
    parents: [
      {
        /** @example '553c2077f0edc3d5dc5d17262f6aa498e69d6f8e'; */
        sha: string;
        /** @example 'https://api.github.com/repos/octocat/Hello-World/commits/553c2077f0edc3d5dc5d17262f6aa498e69d6f8e'; */
        url: string;
      },
      {
        /** @example '762941318ee16e59dabbacb1b4049eec22f0d303'; */
        sha: string;
        /** @example 'https://api.github.com/repos/octocat/Hello-World/commits/762941318ee16e59dabbacb1b4049eec22f0d303'; */
        url: string;
      },
    ];
    /** @example 'https://api.github.com/repos/octocat/Hello-World/commits/7fd1a60b01f91b314f59955a4e4d4e80d8edf11d'; */
    url: string;
    committer: {
      /** @example ''; */
      gravatar_id: string;
      /** @example 'https://secure.gravatar.com/avatar/7ad39074b0584bc555d0417ae3e7d974?d=https://a248.e.akamai.net/assets.github.com%2Fimages%2Fgravatars%2Fgravatar-140.png'; */
      avatar_url: string;
      /** @example 'https://api.github.com/users/octocat'; */
      url: string;
      id: number;
      /** @example 'octocat'; */
      login: string;
    };
  };
  _links: {
    /** @example 'https://github.com/octocat/Hello-World/tree/master'; */
    html: string;
    /** @example 'https://api.github.com/repos/octocat/Hello-World/branches/master'; */
    self: string;
  };
  protected: boolean;
  protection: {
    enabled: boolean;
    required_status_checks: {
      /** @example 'non_admins'; */
      enforcement_level: string;
      /** @example ['ci-test', 'linter'] */
      contexts: string[];
    };
  };
  /** @example 'https://api.github.com/repos/octocat/hello-world/branches/master/protection'; */
  protection_url: string;
}

export interface GhCreateCommitResponse {
  /** @example '7638417db6d59f3c431d3e1f261cc637155684cd'; */
  sha: string;
  /** @example 'MDY6Q29tbWl0NzYzODQxN2RiNmQ1OWYzYzQzMWQzZTFmMjYxY2M2MzcxNTU2ODRjZA=='; */
  node_id: string;
  /** @example 'https://api.github.com/repos/octocat/Hello-World/git/commits/7638417db6d59f3c431d3e1f261cc637155684cd'; */
  url: string;
  author: {
    /** @example '2014-11-07T22:01:45Z'; */
    date: string;
    /** @example 'Monalisa Octocat'; */
    name: string;
    /** @example 'octocat@github.com'; */
    email: string;
  };
  committer: {
    /** @example '2014-11-07T22:01:45Z'; */
    date: string;
    /** @example 'Monalisa Octocat'; */
    name: string;
    /** @example 'octocat@github.com'; */
    email: string;
  };
  /** @example 'my commit message'; */
  message: string;
  tree: {
    /** @example 'https://api.github.com/repos/octocat/Hello-World/git/trees/827efc6d56897b048c772eb4087f854f46256132'; */
    url: string;
    /** @example '827efc6d56897b048c772eb4087f854f46256132'; */
    sha: string;
  };
  parents: [
    {
      /** @example 'https://api.github.com/repos/octocat/Hello-World/git/commits/7d1b31e74ee336d15cbd21741bc88a537ed063a0'; */
      url: string;
      /** @example '7d1b31e74ee336d15cbd21741bc88a537ed063a0'; */
      sha: string;
    },
  ];
  verification: {
    verified: boolean;
    /** @example 'unsigned'; */
    reason: string;
    signature: null | any;
    payload: null | any;
  };
}

export interface GhCreateReferenceResponse {
  /** @example 'refs/heads/featureA'; */
  ref: string;
  /** @example 'MDM6UmVmcmVmcy9oZWFkcy9mZWF0dXJlQQ=='; */
  node_id: string;
  /** @example 'https://api.github.com/repos/octocat/Hello-World/git/refs/heads/featureA'; */
  url: string;
  object: {
    /** @example 'commit'; */
    type: string;
    /** @example 'aa218f56b14c9653891f9e74264a383fa43fefbd'; */
    sha: string;
    /** @example 'https://api.github.com/repos/octocat/Hello-World/git/commits/aa218f56b14c9653891f9e74264a383fa43fefbd'; */
    url: string;
  };
}

export interface GhUpdateReferenceResponse {
  /** @example 'refs/heads/featureA'; */
  ref: string;
  /** @example 'MDM6UmVmcmVmcy9oZWFkcy9mZWF0dXJlQQ=='; */
  node_id: string;
  /** @example 'https://api.github.com/repos/octocat/Hello-World/git/refs/heads/featureA'; */
  url: string;
  object: {
    /** @example 'commit'; */
    type: string;
    /** @example 'aa218f56b14c9653891f9e74264a383fa43fefbd'; */
    sha: string;
    /** @example 'https://api.github.com/repos/octocat/Hello-World/git/commits/aa218f56b14c9653891f9e74264a383fa43fefbd'; */
    url: string;
  };
}
// ...
export interface GhCreateTagObjectResponse {
  /** @example 'MDM6VGFnOTQwYmQzMzYyNDhlZmFlMGY5ZWU1YmM3YjJkNWM5ODU4ODdiMTZhYw=='; */
  node_id: string;
  /** @example 'v0.0.1'; */
  tag: string;
  /** @example '940bd336248efae0f9ee5bc7b2d5c985887b16ac'; */
  sha: string;
  /** @example 'https://api.github.com/repos/octocat/Hello-World/git/tags/940bd336248efae0f9ee5bc7b2d5c985887b16ac'; */
  url: string;
  /** @example 'initial version'; */
  message: string;
  tagger: {
    /** @example 'Monalisa Octocat'; */
    name: string;
    /** @example 'octocat@github.com'; */
    email: string;
    /** @example '2014-11-07T22:01:45Z'; */
    date: string;
  };
  object: {
    /** @example 'commit'; */
    type: string;
    /** @example 'c3d0be41ecbe669545ee3e94d31ed9a4bc91ee3c'; */
    sha: string;
    /** @example 'https://api.github.com/repos/octocat/Hello-World/git/commits/c3d0be41ecbe669545ee3e94d31ed9a4bc91ee3c'; */
    url: string;
  };
  verification: {
    verified: boolean;
    /** @example 'unsigned'; */
    reason: string;
    signature: null | any;
    payload: null | any;
  };
}